requirements.txt
Django==2.2.16
djangorestframework==3.13.1
Pillow==8.3.1
pyyaml==6.0
uritemplate==4.1.1